Restructure misc
