+++ lib/components/sections/card_section.dart
@@ -1,4 +1,4 @@
-import 'package:brunstadtv_app/components/study_topic_card_mini.dart';
+import 'package:brunstadtv_app/components/study/study_topic_card_mini.dart';
 import 'package:brunstadtv_app/graphql/schema/pages.graphql.dart';
 import 'package:brunstadtv_app/helpers/extensions.dart';
 import 'package:brunstadtv_app/models/analytics/sections.dart';
@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 
 import '../../graphql/queries/page.graphql.dart';
 import '../../providers/inherited_data.dart';
-import '../study_topic_card_large.dart';
+import '../study/study_topic_card_large.dart';
 
 class CardSection extends StatelessWidget {
   final Fragment$Section$$CardSection data;
