+++ lib/components/study/study_progress.dart
@@ -4,9 +4,9 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_svg/svg.dart';
-import '../flavors.dart';
-import '../theme/bccm_gradients.dart';
-import '../helpers/ui/svg_icons.dart';
+import '../../flavors.dart';
+import '../../theme/bccm_gradients.dart';
+import '../../helpers/ui/svg_icons.dart';
 
 enum LessonProgressTreeVariant { border2, border3 }
